Move 'accumulators.hpp' out of the 'detail' folder
